feat: add support for providing command context in parameters
This is feature that I experiment with while working on the Firefox Translations training pipeline, and found it very helpful to be able to be able to include parameters in commands without the need for a custom transform.

It's worth calling out that this _does_ begin to introduce more complexity into taskgraph's built-in command context - most notably that there are now three places that it can be provided, which means we need to have a precedence order for them. (I'm using direct > params > file for now, but I could see an argument for parameters being the highest priority.)

If this is something that we want to take in upstream taskgraph I'll update the docs as well. (I realize this might be not be something we want to upstream...so I'm putting that off for the moment.)